tools/etcd-dump-metrics: automate metrics fetch from release binaries
